rename: 일반 코드와 테스트 코드의 네이밍 통일
