+++ moview/modules/question_generator/__init__.py
@@ -1,4 +1,4 @@
 from moview.modules.question_generator.answer_filter import AnswerFilter
 from moview.modules.question_generator.answer_category_classifier import AnswerCategoryClassifier
-from moview.modules.question_generator.answer_subcategory_classifier import AnswerSubCategoryClassifier
+from moview.modules.question_generator.answer_sub_category_classifier import AnswerSubCategoryClassifier
 from moview.modules.question_generator.followup_question_giver import FollowUpQuestionGiver
