+++ tests/modules/input/test_input_analyzer.py
@@ -3,7 +3,7 @@ from tests.common_code_for_test import is_not_none_string
 from moview.modules.input.input_analyzer import InputAnalyzer
 
 
-class TestInitialInputAnalyzer(unittest.TestCase):
+class TestInputAnalyzer(unittest.TestCase):
     def setUp(self) -> None:
         self.analyzer = InputAnalyzer()
